Rubrieken verwijderen moet niet mogelijk zijn wanneer ze gebruikt worden

Closes #700

See merge request commonground/huishoudboekje/app-new!658

+++ backend/hhb_backend/graphql/mutations/rubrieken/delete_rubriek.py
@@ -1,6 +1,8 @@
 """ GraphQL mutation for deleting a Rubriek """
 import os
 
+from numpy import empty
+
 import graphene
 import requests
 from graphql import GraphQLError
@@ -35,6 +37,19 @@ class DeleteRubriek(graphene.Mutation):
         """ Delete current rubriek """
         id = _kwargs.get("id")
         previous = await hhb_dataloader().rubrieken_by_id.load(id)
+
+        # Check if in use by afspraken
+        afspraken = previous.get("afspraken")
+        if afspraken:
+            raise GraphQLError("Rubriek wordt gebruikt in een of meerdere afspraken - verwijderen is niet mogelijk.")
+
+        # Check if in use by journaalposten
+        grootboekrekening_id = previous.get("grootboekrekening_id")
+        if grootboekrekening_id:
+            journaalposten = requests.get(f"{settings.HHB_SERVICES_URL}/journaalposten/?filter_grootboekrekeningen={grootboekrekening_id}").json()['data']
+            if journaalposten:
+                raise GraphQLError("Rubriek zit in grootboekrekening die wordt gebruikt in journaalposten - verwijderen is niet mogelijk.")
+
         delete_response = requests.delete(f"{settings.HHB_SERVICES_URL}/rubrieken/{id}")
         if delete_response.status_code != 204:
             raise GraphQLError(f"Upstream API responded: {delete_response.json()}")
